routing_pattern/assistant_selector/pyproject.toml
[project]
name = "assistant-selector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Abdulqadir8", email = "aq98123@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.0.9",
    "langgraph>=0.2.74",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]

[project.scripts]
run_selector = "assistant_selector.workflow:run_example"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
